+++ dia-batching-server/src/api/error.rs
@@ -10,9 +10,9 @@ pub enum ApiError {
 impl fmt::Display for ApiError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
-            ApiError::CoingeckoError(e) => write!(f, "Coingecko error: {}", e),
-            ApiError::CustomError(e) => write!(f, "Custom error: {}", e),
-            ApiError::PolygonError(e) => write!(f, "Polygon error: {}", e),
+            ApiError::CoingeckoError(e) | ApiError::CustomError(e) | ApiError::PolygonError(e) => {
+                write!(f, "{}", e)
+            }
         }
     }
 }
@@ -24,7 +24,7 @@ impl fmt::Display for CoingeckoError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let CoingeckoError(ref err_msg) = *self;
         // Log the error message
-        log::error!("CoingeckoError: {}", err_msg);
+        log::error!("CoinGeckoError: {}", err_msg);
         // Write the error message to the formatter
         write!(f, "{}", err_msg)
     }
